+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Additions
 
 - the "predefined world spawn position" (set in the server config) now respects Spawn Point Delegation Blocks
+- added "Data Modification Block", has currently one function: add integer values to data when used like a "Data Relay Block"
 
 ## Fixes